Wed, Jul 20, 2022, 10:40 AM +00:00

+++ filter-less/helpers.c
@@ -267,11 +267,11 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
             else
             {
                 // Blue
-                image[i][j].rgbtBlue = round(((float)up.rgbtBlue + down.rgbtBlue + left.rgbtBlue + right.rgbtBlue + up_left.rgbtBlue + up_right.rgbtBlue + down_left.rgbtBlue + down_right.rgbtBlue)/ 8.0);
+                image[i][j].rgbtBlue = round((up.rgbtBlue + down.rgbtBlue + left.rgbtBlue + right.rgbtBlue + up_left.rgbtBlue + up_right.rgbtBlue + down_left.rgbtBlue + down_right.rgbtBlue)/ 8.0);
                 // Green
-                image[i][j].rgbtGreen = round(((float)up.rgbtGreen + down.rgbtGreen + left.rgbtGreen + right.rgbtGreen + up_left.rgbtGreen + up_right.rgbtGreen + down_left.rgbtGreen + down_right.rgbtGreen)/ 8.0);
+                image[i][j].rgbtGreen = round((up.rgbtGreen + down.rgbtGreen + left.rgbtGreen + right.rgbtGreen + up_left.rgbtGreen + up_right.rgbtGreen + down_left.rgbtGreen + down_right.rgbtGreen)/ 8.0);
                 // Red
-                image[i][j].rgbtRed = round(((float)up.rgbtRed + down.rgbtRed + left.rgbtRed + right.rgbtRed + up_left.rgbtRed + up_right.rgbtRed + down_left.rgbtRed + down_right.rgbtRed)/ 8.0);
+                image[i][j].rgbtRed = round((up.rgbtRed + down.rgbtRed + left.rgbtRed + right.rgbtRed + up_left.rgbtRed + up_right.rgbtRed + down_left.rgbtRed + down_right.rgbtRed)/ 8.0);
             }
 
         }
